update/requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
alabaster==0.7.12
algoliasearch==2.6.1
anyascii==0.3.0
anyio==3.6.1
appdirs==1.4.4
appnope==0.1.3
APScheduler==3.6.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.2
attrs==23.2.0
audioread==2.1.9
Automat==20.2.0
Babel==2.14.0
beautifulsoup4==4.11.1
binaryornot==0.4.4
bleach==4.1.0
blis==0.4.1
boto3==1.20.29
botocore==1.23.29
cachetools==4.2.2
catalogue==1.0.0
certifi==2024.8.30
cffi==1.15.0
characteristic==14.3.0
chardet==3.0.4
charset-normalizer==2.0.12
click==8.0.3
cloudpickle==2.0.0
cloudscraper==1.2.60
coloredlogs==15.0.1
comm==0.2.1
ConfigArgParse==1.5.3
constantly==15.1.0
cookiecutter==2.5.0
coqpit==0.0.14
crispy-bootstrap5==2023.10
cycler==0.11.0
cymem==2.0.6
Cython==0.29.26
dateparser==1.0.0
debugpy==1.8.0
decorator==5.1.0
deepspeech==0.9.3
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.4
Django==4.2.9
django-crispy-forms==2.1
dnspython==2.2.1
docopt==0.6.2
docutils==0.17.1
entrypoints==0.3
exceptiongroup==1.2.0
executing==2.0.1
fakeredis==1.7.0
fastapi==0.81.0
fastjsonschema==2.19.1
fbmessenger==4.3.1
filelock==3.4.2
Flask==2.0.2
flatbuffers==2.0
fonttools==4.28.5
fqdn==1.5.1
frozenlist==1.2.0
fsspec==2021.11.1
future==0.18.2
gast==0.4.0
gdown==4.2.0
geographiclib==1.52
geopy==1.18.1
gevent==21.12.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.19.1
greenlet==1.1.2
grpcio==1.43.0
gruut==2.0.4
gruut-ipa==0.10.1
gruut-lang-cs==2.0.0
gruut-lang-de==2.0.0
gruut-lang-es==2.0.0
gruut-lang-fr==2.0.0
gruut-lang-it==2.0.0
gruut-lang-nl==2.0.0
gruut-lang-pt==2.0.0
gruut-lang-ru==2.0.0
gruut-lang-sv==2.0.0
gspread==3.0.1
h5py==3.6.0
httplib2==0.20.2
humanfriendly==10.0
hyperlink==21.0.0
idna==3.8
imagesize==1.3.0
importlib-metadata==4.10.0
incremental==21.3.0
inflect==5.3.0
ipykernel==6.29.0
ipython==8.18.1
ipython-genutils==0.2.0
isoduration==20.11.0
itsdangerous==2.0.1
jedi==0.19.1
jieba==0.42.1
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.1.0
json5==0.9.14
jsonlines==1.2.0
jsonpickle==2.0.0
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.0.11
jupyterlab-pygments==0.1.2
jupyterlab_server==2.25.2
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
klein==21.8.0
libclang==12.0.0
librosa==0.8.0
llvmlite==0.36.0
mailchimp3==3.0.16
Markdown==3.3.6
markdown-it-py==3.0.0
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mecab-python3==1.0.3
mistune==3.0.2
multidict==5.2.0
murmurhash==1.0.6
nbclient==0.5.9
nbconvert==7.14.2
nbformat==5.9.2
nbsphinx==0.8.8
nest-asyncio==1.5.4
networkx==2.6.3
notebook_shim==0.2.3
num2words==0.5.10
numba==0.53.0
numpy==1.19.5
oauth2client==4.1.3
oauthlib==3.1.1
opencv-python==4.8.1.78
opt-einsum==3.3.0
overrides==7.7.0
packaging==21.3
pandas==1.4.4
pandoc==2.0.1
pandocfilters==1.5.0
parso==0.8.3
pathlib==1.0.1
pexpect==4.9.0
Pillow==9.0.0
plac==1.1.3
platformdirs==4.1.0
plumbum==1.7.2
ply==3.11
pooch==1.5.2
preshed==3.0.6
prometheus-client==0.19.0
prompt-toolkit==3.0.43
psutil==5.9.8
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.1
Pygments==2.17.2
pykwalify==1.6.0
pymongo==4.0.2
pynndescent==0.5.5
pyparsing==3.1.4
pypinyin==0.44.0
pyrsistent==0.18.0
pysbd==0.3.4
PySocks==1.7.1
python-crfsuite==0.9.7
python-dateutil==2.9.0.post0
python-dotenv==0.20.0
python-json-logger==2.0.7
python-slugify==8.0.1
python-telegram-bot==13.10
python-ulid==1.1.0
pytz==2024.1
pytz-deprecation-shim==0.1.0.post0
pyworld==0.3.0
PyYAML==6.0
pyzmq==25.1.2
rasa==0.1.1
rasa-core==0.8.6
rasa-nlu==0.11.5
redis==4.0.2
referencing==0.33.0
regex==2021.11.10
requests==2.27.1
requests-oauthlib==1.3.0
requests-toolbelt==1.0.0
resampy==0.2.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.17.1
rsa==4.8
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
s3transfer==0.5.0
scikit-learn==1.0.2
scipy==1.7.3
Send2Trash==1.8.2
simplejson==3.17.6
six==1.16.0
slackclient==2.9.3
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
sounddevice==0.4.4
SoundFile==0.10.3.post1
soupsieve==2.6
sox==1.5.0
spacy==2.2.4
Sphinx==4.3.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.4.36
sqlalchemy2-stubs==0.0.2a27
sqlmodel==0.0.8
sqlparse==0.4.4
srsly==1.0.5
stack-data==0.6.3
starlette==0.19.1
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.4.1
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
terminado==0.18.0
testpath==0.5.0
text-unidecode==1.3
thinc==7.4.0
threadpoolctl==3.0.0
tinycss2==1.2.1
tomli==2.0.1
torch==1.10.1
torchaudio==0.10.1
tornado==6.4
tqdm==4.62.3
traitlets==5.14.1
TTS==0.5.0
Tubes==0.2.0
Twisted==21.7.0
types-python-dateutil==2.8.19.20240106
typing==3.7.4.3
typing_extensions==4.3.0
tzdata==2024.1
tzlocal==4.1
umap-learn==0.5.1
unidic-lite==1.0.8
uri-template==1.3.0
urllib3==1.26.20
virtualenv==20.13.0
wasabi==0.9.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
webrtcvad==2.0.10
websocket-client==1.7.0
Werkzeug==2.0.2
wrapt==1.13.3
yarl==1.7.2
zipp==3.7.0
zope.event==4.5.0
zope.interface==5.4.0
